refactor(kernel): keep the content-object layout helpers in the claims-layout module the offline closure already pins
